Split bootstrap migration tests by behaviour
Move disk-to-database and rename migration tests into focused modules, share their fixtures through a small test-support module, and remove the duplicated disk migration test module.
