+++ src/bootstrap/tests/migration.rs
@@ -1,204 +1,9 @@
-//! Tests for bootstrap migration and upsert helpers.
+//! Tests for bootstrap JSON migration and upsert helpers.
 
-use std::collections::HashMap;
-use std::io::ErrorKind;
-#[cfg(unix)]
-use std::os::unix::fs::PermissionsExt;
-use std::sync::Mutex;
-
-use rstest::{fixture, rstest};
-use tempfile::{TempDir, tempdir};
-use tracing_test::traced_test;
-
-use crate::db::{DbFuture, SettingKey, SettingsStore, UserId};
-use crate::error::DatabaseError;
-use crate::history::SettingRow;
 use crate::testing::test_utils::EnvVarsGuard;
+use tempfile::tempdir;
 
 use super::super::*;
-
-#[derive(Clone, Copy)]
-enum RenameSetup {
-    ExistingFile,
-    MissingFile,
-    #[cfg(unix)]
-    ReadOnlyDirectory,
-}
-
-struct RenameFixture {
-    dir: TempDir,
-    path: std::path::PathBuf,
-    #[cfg(unix)]
-    original_dir_permissions: Option<std::fs::Permissions>,
-}
-
-#[derive(Debug, Default)]
-struct MigrationStoreState {
-    has_settings_calls: usize,
-    set_all_settings_calls: usize,
-    set_setting_calls: usize,
-    captured_settings: HashMap<String, serde_json::Value>,
-}
-
-#[derive(Debug)]
-struct MigrationStore {
-    has_settings_result: Result<bool, &'static str>,
-    set_all_settings_result: Result<(), &'static str>,
-    state: Mutex<MigrationStoreState>,
-}
-
-impl MigrationStore {
-    fn new(has_settings_result: Result<bool, &'static str>) -> Self {
-        Self {
-            has_settings_result,
-            set_all_settings_result: Ok(()),
-            state: Mutex::new(MigrationStoreState::default()),
-        }
-    }
-
-    fn with_set_all_error() -> Self {
-        Self {
-            has_settings_result: Ok(false),
-            set_all_settings_result: Err("forced set_all_settings failure"),
-            state: Mutex::new(MigrationStoreState::default()),
-        }
-    }
-
-    fn state(&self) -> std::sync::MutexGuard<'_, MigrationStoreState> {
-        self.state.lock().expect("migration store state lock")
-    }
-}
-
-impl SettingsStore for MigrationStore {
-    fn get_setting<'a>(
-        &'a self,
-        _user_id: UserId,
-        _key: SettingKey,
-    ) -> DbFuture<'a, Result<Option<serde_json::Value>, DatabaseError>> {
-        Box::pin(async { Ok(None) })
-    }
-
-    fn get_setting_full<'a>(
-        &'a self,
-        _user_id: UserId,
-        _key: SettingKey,
-    ) -> DbFuture<'a, Result<Option<SettingRow>, DatabaseError>> {
-        Box::pin(async { Ok(None) })
-    }
-
-    fn set_setting<'a>(
-        &'a self,
-        _user_id: UserId,
-        _key: SettingKey,
-        _value: &'a serde_json::Value,
-    ) -> DbFuture<'a, Result<(), DatabaseError>> {
-        Box::pin(async {
-            self.state().set_setting_calls += 1;
-            Ok(())
-        })
-    }
-
-    fn delete_setting<'a>(
-        &'a self,
-        _user_id: UserId,
-        _key: SettingKey,
-    ) -> DbFuture<'a, Result<bool, DatabaseError>> {
-        Box::pin(async { Ok(false) })
-    }
-
-    fn list_settings<'a>(
-        &'a self,
-        _user_id: UserId,
-    ) -> DbFuture<'a, Result<Vec<SettingRow>, DatabaseError>> {
-        Box::pin(async { Ok(Vec::new()) })
-    }
-
-    fn get_all_settings<'a>(
-        &'a self,
-        _user_id: UserId,
-    ) -> DbFuture<'a, Result<HashMap<String, serde_json::Value>, DatabaseError>> {
-        Box::pin(async { Ok(HashMap::new()) })
-    }
-
-    fn set_all_settings<'a>(
-        &'a self,
-        _user_id: UserId,
-        settings: &'a HashMap<String, serde_json::Value>,
-    ) -> DbFuture<'a, Result<(), DatabaseError>> {
-        Box::pin(async move {
-            let mut state = self.state();
-            state.set_all_settings_calls += 1;
-            state.captured_settings = settings.clone();
-            drop(state);
-
-            self.set_all_settings_result
-                .map_err(|message| DatabaseError::Query(message.to_string()))
-        })
-    }
-
-    fn has_settings<'a>(&'a self, _user_id: UserId) -> DbFuture<'a, Result<bool, DatabaseError>> {
-        Box::pin(async {
-            self.state().has_settings_calls += 1;
-            self.has_settings_result
-                .map_err(|message| DatabaseError::Query(message.to_string()))
-        })
-    }
-}
-
-impl RenameFixture {
-    fn prepare(&mut self, setup: RenameSetup) {
-        match setup {
-            RenameSetup::ExistingFile => self.write_legacy_file(),
-            RenameSetup::MissingFile => {}
-            #[cfg(unix)]
-            RenameSetup::ReadOnlyDirectory => {
-                self.write_legacy_file();
-                self.make_dir_read_only();
-            }
-        }
-    }
-
-    fn write_legacy_file(&self) {
-        std::fs::write(&self.path, "{}").expect("write legacy settings file");
-    }
-
-    #[cfg(unix)]
-    fn make_dir_read_only(&mut self) {
-        self.original_dir_permissions = Some(
-            std::fs::metadata(self.dir.path())
-                .expect("read directory metadata")
-                .permissions(),
-        );
-        std::fs::set_permissions(self.dir.path(), std::fs::Permissions::from_mode(0o555))
-            .expect("make directory read-only");
-    }
-
-    fn migrated_path(&self) -> std::path::PathBuf {
-        self.dir.path().join("settings.json.migrated")
-    }
-}
-
-#[cfg(unix)]
-impl Drop for RenameFixture {
-    fn drop(&mut self) {
-        if let Some(permissions) = self.original_dir_permissions.take() {
-            std::fs::set_permissions(self.dir.path(), permissions)
-                .expect("restore temp directory permissions");
-        }
-    }
-}
-
-#[fixture]
-fn rename_fixture() -> RenameFixture {
-    let dir = tempdir().expect("create temp dir for rename test");
-    let path = dir.path().join("settings.json");
-    RenameFixture {
-        dir,
-        path,
-        #[cfg(unix)]
-        original_dir_permissions: None,
-    }
-}
 
 #[test]
 fn test_migrate_bootstrap_json_to_env() {
@@ -303,119 +108,6 @@ fn test_libsql_autodetect_does_not_override_explicit_backend() {
     );
 }
 
-fn write_legacy_settings(dir: &TempDir) -> std::path::PathBuf {
-    let settings_path = dir.path().join("settings.json");
-    std::fs::write(
-        &settings_path,
-        serde_json::json!({
-            "onboard_completed": true,
-            "database_backend": "libsql"
-        })
-        .to_string(),
-    )
-    .expect("write legacy settings.json");
-    settings_path
-}
-
-#[tokio::test]
-async fn migrate_disk_to_db_from_dir_missing_legacy_file_is_noop() {
-    let dir = tempdir().expect("create temp dir for missing settings migration");
-    let store = MigrationStore::new(Ok(false));
-
-    super::super::migration::migrate_disk_to_db_from_dir(&store, "test-user", dir.path())
-        .await
-        .expect("missing settings migration should succeed");
-
-    let state = store.state();
-    assert_eq!(state.has_settings_calls, 0);
-    assert_eq!(state.set_all_settings_calls, 0);
-    assert_eq!(state.set_setting_calls, 0);
-    assert!(!dir.path().join("settings.json.migrated").exists());
-}
-
-#[tokio::test]
-async fn migrate_disk_to_db_from_dir_renames_when_db_already_has_settings() {
-    let dir = tempdir().expect("create temp dir for stale settings migration");
-    let settings_path = write_legacy_settings(&dir);
-    let store = MigrationStore::new(Ok(true));
-
-    super::super::migration::migrate_disk_to_db_from_dir(&store, "test-user", dir.path())
-        .await
-        .expect("stale settings migration should succeed");
-
-    let state = store.state();
-    assert_eq!(state.has_settings_calls, 1);
-    assert_eq!(state.set_all_settings_calls, 0);
-    assert_eq!(state.set_setting_calls, 0);
-    assert!(!settings_path.exists());
-    assert!(dir.path().join("settings.json.migrated").exists());
-}
-
-#[tokio::test]
-async fn migrate_disk_to_db_from_dir_writes_settings_and_renames_legacy_file() {
-    let dir = tempdir().expect("create temp dir for settings migration");
-    let settings_path = write_legacy_settings(&dir);
-    let store = MigrationStore::new(Ok(false));
-
-    super::super::migration::migrate_disk_to_db_from_dir(&store, "test-user", dir.path())
-        .await
-        .expect("settings migration should succeed");
-
-    let state = store.state();
-    assert_eq!(state.has_settings_calls, 1);
-    assert_eq!(state.set_all_settings_calls, 1);
-    assert_eq!(state.set_setting_calls, 0);
-    assert_eq!(
-        state.captured_settings.get("onboard_completed"),
-        Some(&serde_json::Value::Bool(true))
-    );
-    assert!(!settings_path.exists());
-    assert!(dir.path().join("settings.json.migrated").exists());
-}
-
-#[tokio::test]
-async fn migrate_disk_to_db_from_dir_db_failure_leaves_legacy_file_unmigrated() {
-    let dir = tempdir().expect("create temp dir for failed settings migration");
-    let settings_path = write_legacy_settings(&dir);
-    let store = MigrationStore::with_set_all_error();
-
-    let error =
-        super::super::migration::migrate_disk_to_db_from_dir(&store, "test-user", dir.path())
-            .await
-            .expect_err("database write failure should abort migration");
-
-    assert!(
-        matches!(error, MigrationError::Database(message) if message.contains("Failed to write settings to DB"))
-    );
-    let state = store.state();
-    assert_eq!(state.has_settings_calls, 1);
-    assert_eq!(state.set_all_settings_calls, 1);
-    assert_eq!(state.set_setting_calls, 0);
-    assert!(settings_path.exists());
-    assert!(!dir.path().join("settings.json.migrated").exists());
-}
-
-#[tokio::test]
-async fn migrate_disk_to_db_from_dir_is_ok_after_best_effort_rename_removed_source() {
-    let dir = tempdir().expect("create temp dir for repeated settings migration");
-    let settings_path = write_legacy_settings(&dir);
-    let store = MigrationStore::new(Ok(false));
-
-    super::super::migration::migrate_disk_to_db_from_dir(&store, "test-user", dir.path())
-        .await
-        .expect("first settings migration should succeed");
-    super::super::migration::migrate_disk_to_db_from_dir(&store, "test-user", dir.path())
-        .await
-        .expect("second settings migration should succeed after source was renamed");
-
-    let state = store.state();
-    assert_eq!(state.has_settings_calls, 1);
-    assert_eq!(state.set_all_settings_calls, 1);
-    assert_eq!(state.set_setting_calls, 0);
-    assert!(!settings_path.exists());
-    assert!(dir.path().join("settings.json.migrated").exists());
-}
-
 #[test]
 fn upsert_bootstrap_vars_creates_file_if_missing() {
     let dir = tempdir().expect("create temp dir for missing-file upsert");
@@ -436,70 +128,6 @@ fn upsert_bootstrap_vars_creates_file_if_missing() {
         parsed[0],
         ("DATABASE_BACKEND".to_string(), "libsql".to_string())
     );
-}
-
-#[traced_test]
-#[rstest]
-#[case::success(RenameSetup::ExistingFile, None)]
-#[case::missing_source(RenameSetup::MissingFile, Some(ErrorKind::NotFound))]
-#[cfg_attr(
-    unix,
-    case::permission_denied(RenameSetup::ReadOnlyDirectory, Some(ErrorKind::PermissionDenied))
-)]
-fn rename_to_migrated_cases(
-    mut rename_fixture: RenameFixture,
-    #[case] setup: RenameSetup,
-    #[case] expected_error_kind: Option<ErrorKind>,
-) {
-    rename_fixture.prepare(setup);
-
-    let result = super::super::migration::rename_to_migrated(&rename_fixture.path);
-
-    match expected_error_kind {
-        Some(kind) => {
-            let error = result.expect_err("rename should fail for this case");
-            assert_eq!(error.kind(), kind);
-            assert!(logs_contain("Failed to rename"));
-        }
-        None => {
-            result.expect("rename legacy settings");
-            assert!(!rename_fixture.path.exists());
-            assert!(rename_fixture.migrated_path().exists());
-            assert!(!logs_contain("Failed to rename"));
-        }
-    }
-}
-
-#[traced_test]
-#[rstest]
-fn rename_legacy_bootstrap_success(mut rename_fixture: RenameFixture) {
-    rename_fixture.path = rename_fixture.dir.path().join("bootstrap.json");
-    rename_fixture.prepare(RenameSetup::ExistingFile);
-
-    super::super::migration::rename_legacy_bootstrap(rename_fixture.dir.path());
-
-    assert!(
-        rename_fixture
-            .dir
-            .path()
-            .join("bootstrap.json.migrated")
-            .exists()
-    );
-    assert!(!logs_contain("Failed to rename"));
-    assert!(logs_contain("Renamed old bootstrap.json to .migrated"));
-}
-
-#[cfg(unix)]
-#[traced_test]
-#[rstest]
-fn rename_legacy_bootstrap_permission_denied(mut rename_fixture: RenameFixture) {
-    rename_fixture.path = rename_fixture.dir.path().join("bootstrap.json");
-    rename_fixture.prepare(RenameSetup::ReadOnlyDirectory);
-
-    super::super::migration::rename_legacy_bootstrap(rename_fixture.dir.path());
-
-    assert!(logs_contain("Failed to rename"));
-    assert!(!logs_contain("Renamed old bootstrap.json to .migrated"));
 }
 
 #[test]
@@ -538,110 +166,4 @@ fn migrate_bootstrap_json_to_env_rename_failure_leaves_env_written() {
         env_path.exists(),
         ".env must still exist after idempotent second run"
     );
-}
-
-mod migrate_disk_to_db_from_dir_tests {
-    use tempfile::tempdir;
-
-    use crate::error::DatabaseError;
-
-    use super::super::super::migration;
-    use super::MigrationStore;
-
-    fn write_legacy_settings(dir: &std::path::Path) {
-        let settings = serde_json::json!({
-            "database_url": "postgres://localhost/ironclaw_test"
-        });
-        std::fs::write(
-            dir.join("settings.json"),
-            serde_json::to_string_pretty(&settings).expect("serialise"),
-        )
-        .expect("write legacy settings.json");
-    }
-
-    #[tokio::test]
-    async fn returns_ok_when_no_legacy_file() {
-        let dir = tempdir().expect("temp dir");
-        let store = MigrationStore::new(Ok(false));
-
-        let result = migration::migrate_disk_to_db_from_dir(&store, "user-1", dir.path()).await;
-
-        assert!(result.is_ok());
-        assert_eq!(store.state().has_settings_calls, 0);
-        assert!(!dir.path().join("settings.json.migrated").exists());
-    }
-
-    #[tokio::test]
-    async fn skips_write_when_db_already_has_settings_and_renames() {
-        let dir = tempdir().expect("temp dir");
-        write_legacy_settings(dir.path());
-
-        let store = MigrationStore::new(Ok(true));
-
-        let result = migration::migrate_disk_to_db_from_dir(&store, "user-1", dir.path()).await;
-
-        assert!(result.is_ok());
-        assert_eq!(store.state().set_all_settings_calls, 0);
-        assert!(!dir.path().join("settings.json").exists());
-        assert!(dir.path().join("settings.json.migrated").exists());
-    }
-
-    #[tokio::test]
-    async fn migrates_and_renames_on_success() {
-        let dir = tempdir().expect("temp dir");
-        write_legacy_settings(dir.path());
-
-        let store = MigrationStore::new(Ok(false));
-
-        let result = migration::migrate_disk_to_db_from_dir(&store, "user-1", dir.path()).await;
-
-        assert!(result.is_ok());
-        assert_eq!(store.state().set_all_settings_calls, 1);
-        assert!(!dir.path().join("settings.json").exists());
-        assert!(dir.path().join("settings.json.migrated").exists());
-    }
-
-    #[tokio::test]
-    async fn returns_migration_error_on_db_failure_and_does_not_rename() {
-        let dir = tempdir().expect("temp dir");
-        write_legacy_settings(dir.path());
-
-        let store = MigrationStore {
-            has_settings_result: Ok(false),
-            set_all_settings_result: Err("injected DB failure"),
-            state: std::sync::Mutex::new(Default::default()),
-        };
-
-        let result = migration::migrate_disk_to_db_from_dir(&store, "user-1", dir.path()).await;
-
-        assert!(result.is_err(), "must propagate DB error");
-        assert!(matches!(
-            store.set_all_settings_result.map_err(|message| DatabaseError::Query(message.to_string())),
-            Err(DatabaseError::Query(message)) if message == "injected DB failure"
-        ));
-        assert!(dir.path().join("settings.json").exists());
-    }
-
-    #[tokio::test]
-    async fn best_effort_rename_on_second_run_returns_ok() {
-        let dir = tempdir().expect("temp dir");
-        write_legacy_settings(dir.path());
-
-        let store = MigrationStore::new(Ok(false));
-
-        migration::migrate_disk_to_db_from_dir(&store, "user-1", dir.path())
-            .await
-            .expect("first run must succeed");
-
-        assert!(!dir.path().join("settings.json").exists());
-
-        let store2 = MigrationStore::new(Ok(false));
-        let result = migration::migrate_disk_to_db_from_dir(&store2, "user-1", dir.path()).await;
-
-        assert!(
-            result.is_ok(),
-            "second run with no settings.json must return Ok"
-        );
-        assert_eq!(store2.state().has_settings_calls, 0);
-    }
 }
