+++ src/bootstrap/tests/mod.rs
@@ -1,4 +1,7 @@
 mod base_dir;
 mod env_format;
 mod migration;
+mod migration_disk_to_db;
+mod migration_rename;
+mod migration_support;
 mod pid_lock;
